+++ src/Hacks/CopyHack.cpp
@@ -5,9 +5,10 @@
 
 using namespace geode::prelude;
 
-class $modify(LevelInfoLayer) {
-    // Copy Hack
-    bool init(GJGameLevel *p0, bool p1) {
+class $modify(LevelInfoLayer)
+{
+    bool init(GJGameLevel *p0, bool p1)
+    {
         if (!LevelInfoLayer::init(p0, p1))
             return false;
 
@@ -22,6 +23,7 @@ class $modify(LevelInfoLayer) {
             m_cloneBtn->getParent()->addChild(aCloneBtn);
             m_cloneBtn->setVisible(false);
         }
+        
         return true;
     }
 };