+++ src/Hacks/MainSongLimit.cpp
@@ -8,21 +8,27 @@ class $modify (SongSelectNode)
 {
     void audioNext(cocos2d::CCObject* p0)
     {
-        if (Client::GetModuleEnabled("default-song-bypass")) {
+        if (Client::GetModuleEnabled("default-song-bypass"))
+        {
             m_selectedSongID++;
             getLevelSettings()->m_level->m_audioTrack = m_selectedSongID;
+
             return SongSelectNode::updateAudioLabel();
         }
+
         SongSelectNode::audioNext(p0);
     }
 
     void audioPrevious(cocos2d::CCObject* p0)
     {
-        if (Client::GetModuleEnabled("default-song-bypass")) {
+        if (Client::GetModuleEnabled("default-song-bypass"))
+        {
             m_selectedSongID--;
             getLevelSettings()->m_level->m_audioTrack = m_selectedSongID;
+            
             return SongSelectNode::updateAudioLabel();
         }
+
         SongSelectNode::audioPrevious(p0);
     }
 };